# Fixed Error in gcp_sql_database if instance is not running Closes #584