--- a/gcp/table_gcp_sql_database.go
+++ b/gcp/table_gcp_sql_database.go
@@ -120,12 +120,14 @@ func listSQLDatabases(ctx context.Context, d *plugin.QueryData, h *plugin.Hydrat
 	instance := h.Item.(*sqladmin.DatabaseInstance)
 
 	// ERROR: rpc error: code = Unknown desc = googleapi: Error 400: Invalid request: Invalid request since instance is not running., invalid
-	// Return nil, if the instance not in running state or in stop state
+	// Return nil, if the instance is not in Runnable state.
 	// ActivationPolicy specifies when the instance is activated.
+        // - SQL_ACTIVATION_POLICY_UNSPECIFIED  Unknown activation plan.
 	// - ALWAYS	The instance is always up and running.
 	// - NEVER	The instance never starts.
-	// - ON_DEMAND The instance starts upon receiving requests.(Deprecated)
+	// - ON_DEMAND  The instance starts upon receiving requests.(Deprecated)
 	// https://cloud.google.com/sql/docs/mysql/admin-api/rest/v1/instances#SqlActivationPolicy
+	
 	if instance.State != "RUNNABLE" || instance.Settings.ActivationPolicy !=  "ALWAYS" {
 		return nil, nil
 	}